gramcnf: Working on BST for bigger expressions

+++ os/np/cmds/gramcnf/tree.h
@@ -29,12 +29,11 @@ struct node_d
     struct node_d *right;
 }; 
 
-
 // Stack
 struct stack_d
 {
-    int top;
-    int items[32];
+    int  top;         // stays int — it's just an index (0..32), no need to widen
+    long items[32];   // was int — this is what actually holds computed values
 };
